// Presentation/cosmeticshop.client/src/apiClient/models/cart-item-request-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * CosmeticShop.WebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

 /**
 * 
 *
 * @export
 * @interface CartItemRequestDto
 */
export interface CartItemRequestDto {

    /**
     * @type {string}
     * @memberof CartItemRequestDto
     */
    productId: string;

    /**
     * @type {number}
     * @memberof CartItemRequestDto
     */
    quantity: number;
}